Address review comments; Move `JavadocTester.copyDir` to `ToolBox`; use FileVisitor code
